Meld special classes with base classes.
This provides the ability to use OrderedGraph or ThinGraph by
simple subclassing of Graph.
