+++ networkx/classes/__init__.py
@@ -2,6 +2,4 @@ from networkx.classes.graph import Graph
 from networkx.classes.digraph import DiGraph
 from networkx.classes.multigraph import MultiGraph
 from networkx.classes.multidigraph import MultiDiGraph
-from networkx.classes.special import SpecialGraph, SpecialDiGraph
-from networkx.classes.special import SpecialMultiGraph, SpecialMultiDiGraph
 from networkx.classes.function import *
